requirements.txt
Flask
python-dotenv
watchdog
Flask-Migrate
psycopg2-binary
Flask-Mail
flask-sqlalchemy
flask_jwt_extended
flask-cors
flask-marshmallow
flask-json
gunicorn
opencv-contrib-python
opencv-python
Pillow
marshmallow
marshmallow-sqlalchemy

#google drive api
google-api-python-client
google-auth-httplib2
google-auth-oauthlib